rename index to recurringIncome, create dashboard files
