+++ views/recurringIncome.handlebars
@@ -1,0 +1,61 @@
+<div class="row center-align">
+<h3>Recuring Income</h3>
+
+<div class="row">
+    <div class="input-field col s6">
+        <select id="employmentStatus">
+            <option value="" disabled selected>Choose your option</option>
+            <option id="1" value="1">Yes</option>
+            <option id="2" value="2">No</option>
+        </select>
+        <label>Are you employed?</label>
+    </div>
+</div>
+
+<div class="row center-align">
+    <div id="recIncName" class="input-field col s6 hidden">
+        <input id="recIncVal" placeholder="" type="text" class="validate">
+        <label for="company_name">What is the name of your company?</label>
+    </div>
+</div>
+
+<div class="row center-align">
+    <div id="paidHourly" class="input-field col s6 hidden">
+        <select id="paidHourlyChoice">
+            <option value="" disabled selected>Choose your option</option>
+            <option value="1">Yes</option>
+            <option value="2">No</option>
+        </select>
+        <label>Are you paid hourly?</label>
+    </div>
+</div>
+
+<div class="row center-align">
+    <div id="annualIncome" class="input-field col s6 hidden">
+        <input id="annualIncVal" placeholder="" type="text" class="validate">
+        <label for="first_name">What is your annual salary?</label>
+    </div>
+</div>
+
+<div class="row center-align">
+    <div id="hourlyIncome" class="input-field col s6 hidden">
+        <input id="hourlyIncVal" placeholder="" type="text" class="validate">
+        <label for="first_name">What is your hourly rate?</label>
+    </div>
+</div>
+
+<div class="row center-align">
+    <div id="hourlyHours" class="input-field col s6 hidden">
+        <input id="hourlyHoursVal" placeholder="" type="text" class="validate">
+        <label for="first_name">How many hours do you work per week?</label>
+    </div>
+</div>
+
+<button id="submitBtn" class="btn waves-effect waves-light" type="submit" name="action">Submit
+    <i class="material-icons right">send</i>
+</button>
+</div>
+
+{{!-- scripts --}}
+<script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
+<script src="/assets/js/recurringIncome.js"></script>
